Disassembler working for single blocks
This adds very preliminary and basic support for disassembling code with
multiple-blocks.

This adds a number of failing tests which demonstrate various
limitations with the block worklist algorithm.

This adds support for recognising data "gaps" in the bytecode after
construction.

Data spliting is now working.  The next outstanding issue is the
handling of JUMPI.  Then I have to implement more opcodes.

This adds a command-line option for disassembly.