Adding abstract stack
This is a new implementation of the abstract stack.  It appears to work
quite well at this stage.  Unfortunately, the code base remains quite
ugly in places and there's a lot of work needed to tidy it up.  However,
at the same time, there is a lot of potential opening up as well.