Update u256 => w256
This changes `u256` to `w256` to reflect the difference in semantics
between native Rust types, and words.  That is, words employ modulo
arithmetic by default (as the EVM does).

For simplicity this now uses two 128bit words for the representation.
However, potentially I might want to change that at some point back to
four 64bit words.  The reason being to reduce the amount of work needed
for managing operations that require a four u64 word representation.