Updating Interval as Abstract Word
This update interval to support traits as necessary to be treated like
an abstract word.

Lots of changeS

It compiles.  Tests are failing though.

Slightly improved interval constructors

Fixes for tests

The key problem was that the stack representation is more precise than
it was previously, and this means the tests failed.