Refactor CfaState => AbstractEvm
This pushes through a fairly significant refactoring which updates the
core EVM abstraction to support both concrete and abstract
representations.  The latter is used for disassembly.

In pushing this through, a number of top-level traits have been
introduced for describing various aspects of a typical dataflow
analysis (e.g. `JoinInto`).  These can be combined together nicely to
manage abstract values.