Add minimal implementation of call/return
This adds a fairly miniaml implementation of call and return statements
to the EvmIL language.  Whilst they do accept arguments, these are not
yet properly implemented.

Also adds a minimal call tests which causes the disassembler to barf.
This was generated from an EIL file using the new call/return
statements.