Execution API is taking shape
The main outstanding question is how to implement bytecodes which have
different behaviours based on concrete values.  For example, when
implementing `LT` in an abstrac state, we need `1` to be placed on the
stack in some cases and `0` in others.  Potentially, we can just have
custom operators which do this.  But, is that enough?  Is there a more
general way?