Remove `asm` module
This removes the `asm` module entirely, and moves it into `bytecode`.
Furthermore, `AssemblyContract` is dropped and replaced simply by
`Assembly`.

At the moment, the static analysis required to refine a legacy contract
is broken.  The problem is that the analysis (+semantics) operate on
byte offsets, whilst contracts have instruction offsets.  For example,
the abstract state operated on uses byte offsets.

Need to think about how best to structure an analysis like this so that
it works.  One obvious option is to allow instructions / semantics to
work on both byte offsets and instruction offsets.