Simplify `Assembly` and `Instruction`
This puts through some fairly dramatic simplifications related to the
mechanism in which instructions are stored within an assembly.  In
particular, there are no longer any labelled instructions (e.g. `PUSHL`,
`LABEL`).  Instead, all instructions are represented in _absolute form_.
To make this work, the assembler and compiler both emit `PUSH2`
bytecodes for all branching instructions (which could easily be
optimised later on).  The key advantage here is that it means an
instruction always has a defined length.