Improve `jumpi` semantics
This puts through an improved semantics for `jumpi` which can exploit
concrete information.  That is, it will not split in the event that the
branch outcome is known.

A key issue is that, in fact, this changed the reachability analysis.
In particular, it meant that bytecode sequences where we have known
conditional branch outcomes were showing unreachable data.  To resolve
this, the reachability analysis now uses a simpler abstract word which
does not do anything very clever.