Implement `havoc` semantics
This puts in place the correct implementation of the `havoc` semantics.
We can now see the `havoc` instruction in action using the `infer`
command-line option.  At this stage, there remain some inefficiencies in
the implementation.  Furthermore, we ideally want some kind of _havoc
analysis_ so that we can test havoc.