Implement dependency analysis
This implements the majority of the dependency analysis, and updates the
cli to support a `--deps` field for printing it out.  That might be
temporary, not sure.

At this stage, however, it remains to actually record the right
information.  The problem is that there is disconnect between the `pc`
value when an item is pushed.