Track correct dependency information
Dependency information is now being tracked correctly.  This was
achieved by adding a `goto()` function onto `EvmStack` and making sure
this was triggered by the enclosing state.  This allows
`DependencyStack` to correctly trace the instruction (byte) offset when
something is pushed on the stack.

At this stage, the encoded information is recorded as _byte offsets_
rather than _instruction offsets_.  This needs to be corrected.