Preliminary havoc analysis is operational!
At this stage, however, there are several outstanding problems.

Firstly, havoc instructions are not actually being executed.  The reason
for this is the trace method which only operates on `pc` values.  Thus,
since `havoc` instructions have no intrinsic length, they are mistakenly
jumped over.

Secondly, havoc instructions are placed for any instruction involved in
a cycle.  That means multiple havoc instructions can be inserted for the
same cycle!

Thirdly, havoc instructions are placed a the example position where the
cycle is detected.  This might be reasonable enough for now, but seems
somewhat suboptimal to me (at least in the long run).