Implement `BlockVec` and `BlockGraph`
These two primitives are the building blocks from which we can extract
the control-flow graph of an instruction sequence.  The code for doing
this is almost all implemented now, though it has not been tested!

Some methods in `BlockGraph` are not yet implemented.  Furthermore, we
would want to add operations for _merging_ and _splitting_ blocks into
`BlockVec`.